Correctly set moderation flag in topics controller
This code wasn't working and this is an indication of the brittleness of
the design (because it depends on the default value of the moderation
flag). Fix the bug and make things more robust by insulating the logic
from the specifics of the default value; in other words, always set the
moderation flag to the appropriate value and save.

Signed-off-by: Wincent Colaiuta <win@wincent.com>