Add comments sweeper
I previously handle comments and topics together in the Topics sweeper
but stumbled across a bizarre bug. On adding a directive like this:

  observe Article, Comment

All named route helpers stopped working in the sweeper, returning nil.
Even non-existent methods would return nil (for example,
"hello_world 1, 2, 3" would return nil rather than raise an exception).

Compare this to the behaviour in a single-model sweeper. If you do a
respond_to? inside such a sweeper you won't see any of the named route
helpers but they do indeed work. Rails (or, perhaps, the Observer
implementation in Ruby) is forwarding the requests for you, and if
you try a "hello_world 1, 2, 3" you'll get an exception as you'd expect.

So evidently some method_missing magic is going on here, and when you
observe multiple models at the same time that magic doesn't know where
to forward your routing helper method requests and so they get dropped
on the floor.

You could manually construct your routes in such a sweeper but that
would be brittle. The solution implemented here is to extract the
comment observation out into a separate sweeper.

Signed-off-by: Wincent Colaiuta <win@wincent.com>