Capistrano: track "master" branch in staging environment
Previously I tracked the "maint" branch in both staging and production
because I wanted production to run the exact same code as had previously
been tested in the staging environment. But in practice it will be
useful to track "master" in the staging environment and "maint" in
production.

In reality the two branches are very close and in fact "master" is
usually just a few commits ahead of "maint"; almost all merges of
"master" into "maint" are simple fast-forwards.

As such it is useful to test the newest code in the staging
environment, and push out urgent fixes to the production environment in
"maint". In normal use, "master" will be tested on the staging
environment and then merged in into "maint" for deployment to the
production environment; in this way the most common case will be for
the production environment to run the exact same code as the staging
environment, which is what we want (we only really want to make
exceptions in exceptional cases, and this commit is all about enabling
us to make that kind of exception when necessary).

Signed-off-by: Wincent Colaiuta <win@wincent.com>