User "post" count should include both comments and topics
Add an "utterances_count" method to the user model which provides the
total number of utterances (currently both comments and topics, but
this can latter be extended to include issues and anything else
appropriate).

This fixes the glitch where a user could register, start a new topic,
reply to his own topic, and with two contributions visible on the same
screen be listed as having "1 post". Now in such a case it will
correctly show "2 posts".

Signed-off-by: Wincent Colaiuta <win@wincent.com>