Move clear_redirection_info up into application controller
This addresses the following scenario: imagine a wiki article, "forums",
which redirects to the real forums path (ie. "/forums"). If the user
clicks on such a link the redirection info is stored in the session and
they get redirected to the forums.

The next wiki page they look at the session info will still be there so
they will see an incorrect "Redirected from forums" label.

The solution is to set up an application-wide after filter which clears
out the redirection info from the session. In this way all controllers
and all actions know how to clear out the info.

The only exception is the articles controller "show" action, because it
is the one that manages and uses the redirection information. We
therefore do a "skip_after_filter" for that action alone, and when it
needs to clear out the information it does so manually.

Signed-off-by: Wincent Colaiuta <win@wincent.com>