Initial version of Bugzilla import script
This is an incomplete-but-working version of the Bugzilla import script.
At this stage we import users, products, bugs and comments. Still to
implement are "CCs" (which will require a new model), and I need to make
a decision on access control seeing as some of the records contain
sensitive information.

Signed-off-by: Wincent Colaiuta <win@wincent.com>