Special handling for space in query strings in paginator
Convert spaces to "+" before link_to sees them because otherwise they'll
get converted into the (ugly) "%20" URL encoded sequence.

This ugliness doesn't usually happen in the view templates because they
tend to use parameter hashes (which convert space to "+"), but they do
happen in the paginator because here we are working with strings
instead.

Signed-off-by: Wincent Colaiuta <win@wincent.com>