Bugzilla import: suppress comments with empty bodies
Not sure what these are, Bugzilla must be using them for some kind of
event, but they clutter up the result so suppress them. There is only
one bug in the entire data set where the original description is empty,
so special case that one with the string "empty" (we don't want to omit
any bugs because we want to preserve the exact same version numbers).

Signed-off-by: Wincent Colaiuta <win@wincent.com>