Extract AJAX "select" invocation into helper method
The "select" invocation is pretty ugly in the view because it is so long
and takes so many parameters (mostly because of the nested
"remote_function" call which itself takes a lot of parameters).

Extract it out into a helper method. I'll be able to use the exact same
method for the other AJAX select controls on this page.

Signed-off-by: Wincent Colaiuta <win@wincent.com>