Fix default_access_options in issues controller
Here I was returning either a Hash or an Array but elsewhere in the code
treated the returned result as a Hash; this bug wasn't noticed earlier
because it only manifests itself for logged-in non-admin users (I had
only tested with admin or anonymous users).

I discovered the issue while expanding the "find_issue" before filter to
provide us with records for "next" and "previous" links, the backend
support for which is already included in this commit.

Signed-off-by: Wincent Colaiuta <win@wincent.com>