Fix incorrect "redirected from" labels
Use the new wiki_redirect? method introduced in 52c004f to eliminate
incorrect "redirected from" labels which could be produced under the
following circumstances:

An article with two wiki links [[foo]] and [[bar]]; the "foo" article
is just an external redirect and the "bar" article is a normal wiki
article. If the user clicks first "foo", then "bar", the articles
controller would inappropriately store "redirected from foo" in the
session, and on seeing "bar" the user would see that message on the
"bar" page even though "foo" and "bar" have no such redirection
relationship.

This change will allow me to jettison the whole after_filter mechanism
in a subsequent commit.

Signed-off-by: Wincent Colaiuta <win@wincent.com>