Cater for anonymous support tickets
I considered a large number of options for how to handle anonymous
support tickets before finally settling on this one. It is a difficult
situation because:

- anonymous users are subject to moderation
- they therefore have no way of seeing their tickets after creating them
- requiring users to sign up first is a barrier to participation
- auto-creating accounts on the fly is complicated and a designing a
  smooth user experience for it is difficult

In the end I decide to do the following:

- make all tickets public by default
- provide user with ability to mark a ticket as private at creation time
- point out to anonymous users that they can only have public tickets
- show "awaiting moderation" on trying to view such a ticket
- provide an Atom feed so that users can watch when tickets are updated

There is still one minor glitch to address: that when a comment is
moderated as ham the cache sweepers do not fire, so the feeds do not
necessarily get updated; I still need to come up with an elegant
solution for that problem.

Signed-off-by: Wincent Colaiuta <win@wincent.com>