Fix view spec failures for products#index
I must admit to seriously questioning the value of mock-heavy view specs
like this. They are brittle, they break whenever you evolve the code
(not merely retrospective changes, but forward-looking additions as
well), and they represent a whole lot of "busy work" which I am not sure
is justified.

So I've fixed the breakage but I think I need a new way of approaching
these view tests in the future.

Signed-off-by: Wincent Colaiuta <win@wincent.com>