Capistrano: fix specs
I have to explicitly use "current_path" here rather than "release_path"
otherwise the the spec task might fail if enough time has elapsed
between the "cap deploy:update" and "cap spec" (because the timestamp
will be different and the directory names won't match).

Note that it is not so easy to run the spec task automatically now
because the "current_path" will actually point to the old version of
the app if executed before "deploy:symlink". We could changed it to
"release_path" but then it wouldn't work if executed as a stand-alone
task. (To solve this conundrum I really need to break this up into two
tasks).

Signed-off-by: Wincent Colaiuta <win@wincent.com>