Fix dead routing helpers in sweepers
The behaviour of Rails' sweepers gets even more mysterious; the
"issue_path" route helper stopped working in the issue sweeper.
As previously noted, there's "method_missing" magic in here which
catches the message send and returns nil. Previously I had thought
this to only affect sweepers which try to observe more than one model
class (my guess was that the observer mechanism didn't know where to
forward the message).

Well, the problem suddenly sprang up in the issue sweeper which only
observes a single model and is basically identical to the article
sweeper, for instance, which does work.

Most puzzling of all this definitely worked yesterday.

So take the magic out of the equation and just include the route helper
methods explicitly. Half of the challenge was finding out how to do
this.

Signed-off-by: Wincent Colaiuta <win@wincent.com>