Sessions controller, redirect to dashboard on login
At the same time, we fix a bug werein a blank "original_uri" would cause
us to redirect to the home path (because '' is non-nil we would redirect
to it; what was needed was a test for "blank-ness" rather than
"nil-ness").

Add some specs to make sure this doesn't happen again.

Signed-off-by: Wincent Colaiuta <win@wincent.com>