Teach paginator to preserve nested hash parameters
I needed this when adding pagination to the issue tracker search results
page. Because the search form uses a nested hash to transmit the search
criteria the old method used here wouldn't work.

In solving the problem I discovered the handy "to_query" method which
allows me to not only handle nested hashes, but also replace the manual
space-to-plus-sign conversion that I was doing, because "to_query" does
this automatically.

Signed-off-by: Wincent Colaiuta <win@wincent.com>