Proof Issues#search action against refreshing
If the user clicks in the location bar and refreshes the page after the
initial POST request the URL path will be /issues/search and the
params[:issue] will be nil. Guard against this possibility to prevent an
500 error (we don't want the user to be able to trigger such an error so
trivially). In this case all issues are shown (at least, all the issues
for which the user has access rights).

If the user clicks any of the pagination or table sorting links then
there is no longer a risk of such an error because the page will have
been retrieved via a GET request and the URL will therefore contain all
the necessary params.

Signed-off-by: Wincent Colaiuta <win@wincent.com>