Inset content inside rounded gray border box
This is partly a concession to the "Web 2.0" aesthetic which valorizes
sparse layouts with lots of whitespace around the content. It does
result in a reduction of the available space on the page but I think
the result is visually more pleasing. Note that I haven't gone so far
as to inset the content by a huge amount and impose a fixed width.
Also note that the rounded corners will only work in WebKit and Gecko.

Signed-off-by: Wincent Colaiuta <win@wincent.com>