Add "major" class to a number of different headings
Minor visual cleanup: make some of the major headings stand out by
adding the "major" class.

While the normal progression of "h1" through to "h6" is freely used
inside things like wiki articles it is useful to have another
"namespace" in which we can insert independent headings that divide
the page at a higher level.

Note that in order to do this I had to drop the use of the "page_title"
method. I added some accompanying specs. I suppose I could add an
optional "major" parameter to the method to replace this explicit
assignment to the "@page_title" instance variable.

Signed-off-by: Wincent Colaiuta <win@wincent.com>