Thin Issues#search action by extracting into model
We do this in the interests of testability, as the thinner controller is
easier to test, and the functionality that was in the controller is more
exposed (and therefore more testable) in the model.

Another hint that this move is the right one is that we no longer have
to use the "send" method to access the private
"sanitize_sql_for_conditions" model method from the controller.

Signed-off-by: Wincent Colaiuta <win@wincent.com>