Add template for Search#issues action
This is a very simple template which just calls the issues/search
partial. It is intended as a landing page for users who click the
"Search" button on the Support page, and also as a fallback target link
for people browsing without JavaScript turned on (does anybody actually
do that nowadays?).

Signed-off-by: Wincent Colaiuta <win@wincent.com>