Rename url_for_comment for comment method
The only reason this method exists is because the comment_path method
dynamically created by Rails doesn't give us the desired paths because
the comment model is not only polymorphic, but sometimes acts as a
doubly nested resource (forums -> topics -> comments) and other times
not (issues -> comments).

Pick a more consistent name for the method, polymorphic_comment_path.

Signed-off-by: Wincent Colaiuta <win@wincent.com>