Read in application config from a YAML file at startup
I want to clean up my mailers a bit and this is the first step. We need
a place were we can store application config in a way that plays nicely
with SCM and with running the same code on two different hosts with
different hostnames and the like.

This problem is already solved by the database.yml file: we can store a
sample in SCM and tailor on a per-machine basis, and use Capistrano to
symlink to the file as part of deployment.

So we adopt a similar strategy, but this time with an app_config.yml
file.

Signed-off-by: Wincent Colaiuta <win@wincent.com>