Simplify mailers by using application config
Here we make use of the application config introduced in c438f52 to
simplify the mailers. Thanks to the config we no longer need to pass
in a request object as a parameter. This in turn opens up the door
towards more mailers, and mailers which aren't called from controllers
(originating some mails from within models may be useful and help keep
controllers thinner).

Signed-off-by: Wincent Colaiuta <win@wincent.com>