Fix "last post" information for anonymous comments
This fixes http://rails.wincent.com/issues/671.

The problem was that if we had a topic with an anonymous reply then the
SELECT query in the find_topics_for_forum method would end up falling
back on the topic owner due to the IFNULL logic.

The solution is easy -- drop the IFNULL logic -- and in this way when we
do the LEFT OUTER JOIN the NULL values from the right hand table
(corresponding to the anonymous user) get carried through to the results
table as NULL values, which is what we want because they are handled in
the view as "anonymous" user comments.

But while dropping the IFNULL logic makes for a simpler query and one
without any MySQL-specific SQL, it introduces another problem. Now when
we have a topic with no comments we no longer show the topic originator
as the "last poster" (this was the whole point of the IFNULL logic in
the first place).

So we make sure that when we create a new topic (which by definition
will have no comments at that time) we populate the "last commenter"
fields apropriately from inside the "before_create" filter.

The other side of the equation must be handled in the "after destroy"
callbacks in the Comment model; what happens when we delete all the
comments? We have to make sure that the "last commenter" info is set
again in that case.

Note that not all commentable classes have "user" attribtues (weblog
posts, for example, and articles too) so I don't try blindly setting
the user attribute as a fallback value.

Lots of specs included to make sure this issue doesn't regress.

Signed-off-by: Wincent Colaiuta <win@wincent.com>