Update comment on find_topics_for_forum method
While tempting to move this into the forum model doing so would be
counterproductive because we'd lose some magic mapping provided by Rails
based on the column names and types that it knows from the schema. I
realized this when I actually tried to move it. Amend the comment to
remind me not to try this in the future.

Signed-off-by: Wincent Colaiuta <win@wincent.com>