Add migration to update "last post" info for topics
Commit 753e6f2 fixed a bug with the display of "last post" information
in the forums#show action. From here on all new topics should have the
correct information. This commit is a migration that goes back and
updates old posts so that they have the correct information as well.
This is not an essential update as any new post to an old topic would
also cause the information to be refreshed, but it's nice to go back
and correct the old posts so that even dead topics which are unlikely
to ever be updated again will have the correct information.

Signed-off-by: Wincent Colaiuta <win@wincent.com>