Workaround Safari rendering of ultra-wide pre blocks
I like to use the "overflow: auto" property on pre blocks to ensure that
they don't make pages unreadable when they have a lot of content (they
can make pages too wide and so even to read the rest of the page content
you have to do a lot of uncomfortable horizontal scrolling).

Seems that if you place a pre block inside a table in Safari then all
bets are off: the table grows as large as is necessary to accommodate
the pre block, thus trading a horizontal scrollbar in a small portion
of the page for one which affects the entire page.

Add a "max-width" property to preven Safari from spoiling the users'
experience, but apply it only to pre blocks inside tables to minimize
the number of cases in which the kludge must be applied. I'd like to
find an alternative solution, so hopefully this will only be temporary.

Signed-off-by: Wincent Colaiuta <win@wincent.com>