Add links to Comments#edit template
These are various links for administrator convenience: destroy, mark
as spam, mark as ham, return to list of comments awaiting moderation
and return to admin dashboard.

These links are especially useful in the light of the "new comment
alert" emails that get sent to the administrator: the alerts contain a
link back to the Comments#edit URL and so it makes sense to provide
buttons on that page for moderating them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>