Make Comments#update redirect differently depending on moderation status
If the edited comment is awaiting moderation then redirect back to the
list of comments awaiting moderation; otherwise, redirect to the path of
the comment itself.

This is a workflow polish issue: if the admin edits a comment that is
awaiting moderation then he most certainly doesn't want to be redirected
to the path of the comment itself (because it won't even be visible when
he gets there).

Signed-off-by: Wincent Colaiuta <win@wincent.com>