Add topic and issue mailers
These are extremely similar to the existing comment mailer. Could think
about inserting a abstract parent class and inheriting from that for all
three mailers, but I think that given the size of the mailers and the
templates, it's actually best to tolerate this small amount of
repetition in the name of clarity.

Signed-off-by: Wincent Colaiuta <win@wincent.com>