Prevent Haml auto-indentation from mangling form text areas
This continues to be my number-one complaint with Haml: that its
auto-indentation wreaks havoc with whitespace-sensntive elements
like text areas. Haml provides itself on the attractive appearance of
its output, but the truth is that the whitespace-preservation hack
produces horrid, ghastly looking HTML.

In the case of this commit we're not just talking about an aesthetic
problem (as is the case with the output from my wikitext translator,
for example, which gets converted into an ugly, punctuation-studded
string by Haml) but an actual data-mutation problem because Haml
destructively mauls the contents of the form.

I'm hoping that the next release of Haml (with the ":ugly" option,
in my opinion not so ugly) will be stable enough to allow me to
jettison this kind of hack.

Signed-off-by: Wincent Colaiuta <win@wincent.com>