Fulltext search updates
Commit 79fe477 upated fulltext search indexing to make use of the
wikitext module. Now we make use of it to tokenize search query
strings prior.

I also include some minor tweaks that correct breakage in the
old search code (invalid or ambiguous SQL).

Now that I have some decently sized test data in the database it
looks like "AND" queries are going to be prohibitively slow; as
you add additional search terms the query time grows exponentially:
(0.1 seconds for 1 word, 0.37 for 2, 0.19 for 3, 2.29 for 4, 6.53
for 5 -- obviously those results are not scientific).

On the other hand, "OR" queries look to be acceptable (0.2 seconds
for a 5-word query).

Signed-off-by: Wincent Colaiuta <win@wincent.com>