Simplify "OR" fulltext search query
Perhaps because the "AND" case uses complicated JOIN logic, I
misguidedly used complicated UNION logic in the "OR" case. This was
completely unnecessary and the query is much simpler and faster if
we just use a simple WHERE clause with the appropriate subclasses
joined together using AND and OR. As a bonus, the code to construct
the query has been made more straightforward as well.

I'd like to be able to pull off a similar simplication for the "AND"
case but I fear that it won't be possible.

Signed-off-by: Wincent Colaiuta <win@wincent.com>