Turn on full-text indexing for articles, issues and topics
Leaving out comments for now because these are somewhat of a thorny
issue: not only can individual comments have ownership and be public or
private, they can also be attached to models which themselves can be
public or private. In other words, a "public" comment could be attached
to a "private" issue, and so should be excluded from search results.

Signed-off-by: Wincent Colaiuta <win@wincent.com>