Add support for pagination in needle model
Try to fetch 20 results at a time, and accept an optional "offset"
parameter so that the caller can fetch subsequent pages. Note that we
actually try to retrieve 21 rows even though we only intend to display
20; this is so that we can detect when there are still more results
available and should display a "more" link.

Seeing as full-text search queries may be expensive we use this trick
rather than performing a separate query beforehand to get the full count
of results.

Signed-off-by: Wincent Colaiuta <win@wincent.com>