Teach "needle" model to prefetch models from search results
The basic "find_using_query_string" method would return an array of
"needle" objects describing the search results in terms of class and id.
In order to actually do anything useful with this information one has to
hit the database again.

Mitigate this typical "N + 1" situation by grouping the results
according to model type and prefetching the actual model instances. The
worst case scenario in the current set-up will be 5 queries (1 for the
search and 4 more to prefetch matching articles, issues, posts and
topics) instead of 22 (1 for the search and 21 for the search results).
The best case scenerio is 2 queries (1 for the search and 1 for
prefetching, where only one type of model is in the search results) or
even 1 query (when there are no hits).

Signed-off-by: Wincent Colaiuta <win@wincent.com>