Topics controller: special case forumless params
Topics are a nested resource inside forums, so up until now I've
required URLs like "/forums/name-of-forum/topics/123". But in the case
of the full-text search this design decision makes it hard to avoid
yet another "N + 1 SELECT" situation, because in order to provide links
to the topics we need to know not only the id of the parent forum but
also its display name.

Various alternatives are available:

- put up with the "N + 1 SELECT" hit (rejected due to excessive
  performance penalty)
- preprocess search results and try to grab all of the associated forums
  in one hit (rejected due to complexity)
- use the :include functionality to do eager-loading of the forums
  (rejected due to heavyweight query in an already performance-critical
  section, the ugliness of the query itself, and the inelegance and
  complexity of special-casing nested resources inside the full-text
  search backend)

So this leaves us with only one more alternative, the one implemented in
this commit. We add a small amount of special case logic to the topics
controller to teach it how to handle "forumless" topics. Basically, for
a request which is missing its forum parameter we look up the forum and
do a redirect to the proper, nested URL.

This solution is the winner because it is not complex, and its
performance profile is good: we don't do redundant queries, we don't
needlessly eager-load resources which may never be needed, and we don't
increase the complexity or resource-intensitivity of the existing
full-text queries; instead, we do one simple additional database lookup,
but only "on demand" (ie. when the user actually clicks on a link).

This pattern simplifies things considerably, and I am thinking about
employing it elsewhere in the app too (for comments, for example, which
apart from being nested are also polymorphic to boot which makes them
awfully cumbersome to deal with in many situations).

Signed-off-by: Wincent Colaiuta <win@wincent.com>