Extend "classifiable" specs for full-text indexing
These specs ensure that the "moderate_as_ham!" and "moderate_as_spam!"
methods are updating the full-text search index when appropriate.

My initial cut at this spec used "should_receive" to make sure that the
"update_needles" method was being called, but the specs failed. It looks
like you can't set up that kind of expectation for private methods, so
I've had to do some set-up and break-down and compare the before, during
and after states to confirm the behaviour instead.

Signed-off-by: Wincent Colaiuta <win@wincent.com>