Add notes on cause of spurious RSpec failure
The following test program demonstrates the problem. Basically we have
code which checks for the existence of a private method at runtime and
runs it if one exists; we can't use "respond_to?" because that will
always return false for private methods, nor can we use "methods"
because that also excludes private methods.

What happens is that RSpec comes along and removes the private method
and replaces it with a public one when setting up the mock; this in
turn means that the "private_methods" check will fail and the spec is
guaranteed to fail also.

Workarounds include:

- patching RSpec to stub/mock methods as private/protected/public
  according to what the stubbed/mocked method is (if it exists)
- adding more conditional logic to the implementation to try both
  "private_methods" _and_ "respond_to?" (but this seems a horrible
  kludge)
- dropping the check from the implementation and instead just trying
  to run the method anyway, catching NameError or perhaps NoMethodError
  (I don't really like this because it could spuriously mask those
  same exceptions if they ever occur _inside_ the called private method)

So for now I think I'm just going to keep my ugly, mockless specs.

require 'rubygems'
require 'spec'

class Thing
  def foo
    if private_methods.include?('bar')
      bar
    else
      nil
    end
  end
private
  def bar
    puts "bar"
  end
end

describe Thing, "foo method" do
  it 'should call bar' do
    thing = Thing.new
    puts thing.private_methods.include?('bar') # true
    puts thing.methods.include?('bar')         # false
    thing.should_receive(:bar)
    puts thing.private_methods.include?('bar') # false
    puts thing.methods.include?('bar')         # true
    thing.foo
  end
end

Signed-off-by: Wincent Colaiuta <win@wincent.com>