Handle deleted and destroyed models
If we do a "model.destroy" then the callbacks are fired and the search
index gets updated, but if we do a "delete" then no callbacks fire and
the search index will have stale information in it.

Prepare the "prefetch" method to handle this eventuality gracefully
instead of raising a RecordNotFound exception. We return nil
placeholders for any missing records, and the view is prepared to
skip over such nil placeholders; in this way pagination isn't broken by
returning fewer records than expected.

Signed-off-by: Wincent Colaiuta <win@wincent.com>