Import styles into maintenance page
The CSS for the maintenance page was broken because of the way nginx is
configured to rewrite _all_ URLs when the maintenance page is up. This
means that it's not possible to include a separate CSS file or any
graphics.

So we move the CSS into the maintenance page itself and simplify so that
not images are referenced.

Signed-off-by: Wincent Colaiuta <win@wincent.com>