Don't auto-preserve when in ugly mode
In ugly mode there is no need to automatically preserve whitespace for
pre and textarea tags because all output will be left-indented anyway.
This means less work for ugly mode and ironically makes the output less
"ugly" because it doesn't insert the HTML entities for newlines.

This is a manual cherry pick of commit 6213749 from the official Haml
repo as I don't want to wait for 2.0.1/2.1 for this change. The ugliness
of the Haml output and the way it massacres the nicely formatted output
of the Wikitext translator has long been my only real complaint about
Haml.

Signed-off-by: Wincent Colaiuta <win@wincent.com>