Fix subtle bugs in partials specs
Turns out that I wasn't doing things in the partials specs the "right"
way, and that in turn meant that my specs which had been passing for
weeks and months on multiple machines were suddenly failing when run
in batch mode ("rake spec", "spec spec") but never when run
individually.

So I've learnt that mocks and stubs on partials are generally off-limits
in partials specs, at least for the purposes of faking local variables.
The right way to do this is to instead just do a "render :partial" call
in the first place; perhaps this should have been obvious to me but at
times one doesn't see the wood for the trees...

Signed-off-by: Wincent Colaiuta <win@wincent.com>