Fixes for Rails gems dependencies
These fixes are needed to address two problems with the current
implementation of gem dependencies in Rails 2.1.0_RC1: that missing gems
can cause "rake gems" to barf with an exception and that frozen gems
with compiled extensions in an "ext" subdirectory aren't properly
handled.

These have been submitted upstream and will hopefully be included in
Rails 2.1 final. See:

  http://rails.lighthouseapp.com/projects/8994/tickets/268

Signed-off-by: Wincent Colaiuta <win@wincent.com>