Fix Rails load ordering of frozen gems
If a user freezes version "X" of a gem into their vendor directory
but a version "Y" of the same gem already exists on the system then we
were incorrectly loading version "Y". The fix is to prepend the frozen
gem's path rather than appending it to ensure that it gets loaded
instead.

This patch has been submitted upstream as well:

  http://rails.lighthouseapp.com/projects/8994/tickets/268

Signed-off-by: Wincent Colaiuta <win@wincent.com>