Clean temporary build files from frozen wikitext gem
When freezing ("rake gems:unpack") and building ("rake gems:build") on
multiple platforms it is a good idea to get rid of temporary build
products to avoid confusion. For example, if you build first on Mac OS
X and then on Linux you don't want the object files floating around
when you do your "rake gems:build" because they will be rejected by
the Linux linker. But neither do you want to do a "make clean" before
committing because that will blow away the extension as well as the
intermediate object files.

The method I've used this first time around is:

  # on Mac OS X
  rake gems:unpack
  rake gems:build
  git add/git commit/git push etc

  # on Linux
  git pull
  cd vendor/gems/wikitext-1.1.1/ext
  ruby extconf.rb
  make clean
  cd ../../../..
  rake gems:build
  git add vendor/gems/wikitext-1.1.1/ext/wikitext.so \
          vendor/gems/wikitext-1.1.1/bin/wikitext.so
  git commit/git push etc

  # back on Mac OS X (preparing for this commit)
  rm vendor/gems/wikitext-1.1.1/ext/*.log \
     vendor/gems/wikitext-1.1.1/ext/*.o

I think that in the future I'll be adding a new task to the wikitext
Makefile to automate this cleanup.

Signed-off-by: Wincent Colaiuta <win@wincent.com>