Don't let spam comments bump issues up the list
Previously we updated the caches in the "commentable" model whenever a
new comment was created. This meant that even a spam comment, even
though invisible to the public while awaiting moderation, could
inappropriately bump an old issue to the top of the "recent issues"
listing.

The fix is to defer updating the caches on comment creation. Only when
the comment is marked as ham will the caches actually get updated.

A subsequent fix will handle the (admittedly unlikely) case in which a
ham comment gets marked as spam, in which case the caches will have to
be revised in the opposite direction.

See: http://rails.wincent.com/issues/1118

Signed-off-by: Wincent Colaiuta <win@wincent.com>