Better cache updating for moderated comments
Now we handle comment moderation in a more general fashion rather than
just the specific case of a comment going from "awaiting moderation" to
"not awaiting moderation/classified as ham"; we actually count the
number of moderated ham comments and get the information about the last
commenter on the fly.

This ensures that we also handle cases like ham comments later being
re-classified as spam comments and so on. It also correctly handles
comments being moderated out of order (previously, if two comments were
moderated as ham then the last-moderated would end up being identified
as the "last comment" even if it were actually posted first).

This is an expensive operation involving a few SELECTs and a couple of
UPDATE statments, but seeing as there is only one admin and only the
admin can moderate this is unlikely to ever become a bottleneck.

Signed-off-by: Wincent Colaiuta <win@wincent.com>