Move Haml back into vendor/plugins (again)
During test deployment to the staging environment I discovered that Haml
(2.0.2) still doesn't work properly when installed in vendor/gems (as of
Rails 2.1.1).

Specifically I was seeing:

  uninitialized constant Rails::Initializer::Sass

Which was preventing migrations, specs, the script/console, or the app
itself from running.

This only happened on the server, which has only the Haml 1.8.2 gem
installed system-wide. On a hunch, I tried uninstalling all system-wide
versions of the Haml gem on my development machine and was able to
reproduce the error.

Haven't yet determined why Haml isn't playing nicely here but it may be
an interaction with a Rails bug, seeing as neither init.rb nor
rails/init.rb are being evaluated when the gem is frozen into
vendor/gems/. When installed in vendor/plugins/, the init.rb file does
indeed get evaluated.

Even so, it does look to be a Haml-specific issue as the same problem
does not appear in the Wikitext gem (again, frozen into vendor/gems/ and
no version install system-wide).

Signed-off-by: Wincent Colaiuta <win@wincent.com>