Update comments about Rails ticket #324
I've looked at the Rails code for this issue and it is pretty hard to
decipher. Not really clear what the fix is but it looks as if only
gems which explicitly included specifications get the special treatment.
(Rails creates specs on the fly for ActiveRecord and friends and that is
probably the solution for vendored gems as well, if it could be done
automatically.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>