Use Rails 2.1 conditional page caching
Here we drop a couple of controller methods used for caching Atom feeds
and replace them with ":condition" Procs passed to the "caches_page"
method.

On the one hand we are making things slightly more complicated as the
condition Proc reads a little less like English.

But on the other hand this is a clear improvement for a number of
reasons:

- removes clutter from the ApplicationController
- makes use of "Rails way" of doing conditional page caching rather
  than using custom methods
- makes exactly what is happening less opaque; that is, the totality of
  the behaviour is now visible to anyone inspecting the controller (it
  is no longer necessary to look at the application controller in order
  to see exactly what's happening, nor is it necessary to know that one
  has to look there to find the "cache_feed" method)

My only remaining gripe with the implementation is that it relies on
"send" in order to access a protected method on the controller. This
would be the case even if we ditched our customer "is_atom?" method and
used "controller.format.atom?" as provided by Rails. This seems to be a
necessary evil of working with the passed-in controller instance from
inside a Proc.

Signed-off-by: Wincent Colaiuta <win@wincent.com>