Update comments in the custom Atom feed helper
When these comments were originally written the ability to override URLs
was necessary in order to work around a bug in the original Rails
version of the plug-in.

Since then the problem went away and "polymorphic_url" could be used in
those cases.

Nevertheless, I later made use of the utility of being able to provide
custom URL overrides for Atom entries; for example, being able to append
an anchor like "#comment_3205" to the end of a URL.

This commit updates the comments to reflect the new state of affairs.

Signed-off-by: Wincent Colaiuta <win@wincent.com>