Fix broken articles feed when no articles
When no articles were in the database we would bomb here trying to
call "updated_at" on nil. Explicitly check for the no articles case and
fall back to using the "Rails Epoch" (Rails 1.0.0 release date, 13
December 2005).

We want to pick a stable date as a default so as not to confuse
newsreaders by constantly providing "Time.now" whenever they query the
feed, and the UNIX Epoch is probably a little too far back in the past
to be appropriate.

Signed-off-by: Wincent Colaiuta <win@wincent.com>