capistrano: warn if not on expected branch
It's all too easy to misdeploy by forgetting to do a "git push" before
running "cap deploy:update"; we already added a reminder to do this
whenever the local repo was ahead of the remote one, but it doesn't
necessarily help when you are not on the right branch.

In other words, you might be working on your "master" (staging) branch
and do a "git push", and then do a production deployment ("maint"
branch). The fact that the "master" branch is up-to-date tells you
nothing about whether the branch you really care about (in this case
"maint" because it's the one you're about to deploy) is up-to-date.

So here we try to warn about this situation by alerting the user
whenever he or she is not on the branch that he is about to deploy.
This warning would have saved me from a couple of recent botched
deployments which failed because I had been working on "master" and
expecting my changes to appear in the production environment (in
other words, I'd forgotten to merge them over to "maint").

Signed-off-by: Wincent Colaiuta <win@wincent.com>